splitting config
