+++ app/client/src/components/designSystems/appsmith/help/DocumentationSearch.tsx
@@ -14,7 +14,7 @@ import styled from "styled-components";
 import { HelpIcons } from "icons/HelpIcons";
 import { HelpBaseURL } from "constants/HelpConstants";
 import { getDefaultRefinement } from "selectors/helpSelectors";
-import { getAppsmithConfigs } from "configs";
+import { getAppsmithConfigs } from "@appsmith/configs";
 import { AppState } from "reducers";
 import {
   setHelpDefaultRefinement,
