+++ app/client/src/ee/configs/index.ts
@@ -1,0 +1,1 @@
+export * from "ce/configs/index";
