+++ app/client/src/ee/configs/types.ts
@@ -1,0 +1,1 @@
+export * from "ce/configs/types";
